Move FileDialogState to Dialog file
